AWS S3 연결 및 리팩토링
