+++ app/model_utils.py
@@ -1,0 +1,13 @@
+import tensorflow as tf
+from app.config import MODEL_PATH
+
+class ModelManager:
+    def __init__(self):
+        self.model = None
+
+    def load_model(self):
+        if self.model is None:
+            self.model = tf.keras.models.load_model(MODEL_PATH)
+        return self.model
+
+model_manager = ModelManager()
